requirements.txt
alembic==1.7.6
appdirs==1.4.3

art==5.4
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.4
colour==0.1.5
cryptography==2.8
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0

dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-cors-headers==3.11.0
django-heroku==0.3.1
django-location-field==2.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
dominate==2.6.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.6
Markdown==3.1.1
MarkupSafe==2.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pbr==5.9.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.3
pycairo==1.16.2
pycodestyle==2.8.0

Pygments==2.3.1

PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-dateutil==2.7.3
python-decouple==3.6
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simple-colors==0.1.5
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
toml==0.10.2

urllib3==1.25.8

virtualenv==20.0.17
visitor==0.1.3
wadllib==1.3.3
Werkzeug==2.0.3
whitenoise==6.0.0
WTForms==3.0.1

zipp==3.7.0
